% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIREDAM.R
\docType{data}
\name{FIREDAM}
\alias{FIREDAM}
\title{Firedam data}
\format{
## `FIREDAM`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Total distance fire spread}
  \item{DAMAGE}{Amount of damage caused by fire}
  ...
}
}
\source{
textbook
}
\usage{
FIREDAM
}
\description{
A collection of fire damage data
}
\keyword{datasets}
