% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myquad.R
\name{mybin}
\alias{mybin}
\title{A binary probability function}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{a numeric vector}

\item{n}{the number to choose from}

\item{p}{the probability of choosing}
}
\value{
a barplot
}
\description{
A binary probability function
}
\examples{
mybin(iter = 10, n = 5, p =0.5)
}
