% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{A bootstrap ci function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{a numeric vector}

\item{x}{x}

\item{fun}{the statistic to calculate}

\item{alpha}{the bounds for the confidence interval}

\item{cx}{cx}

\item{...}{stuff}
}
\value{
a plot
}
\description{
A bootstrap ci function
}
\examples{
myboot2(x=c(1,1,1,2,2,2,3,3,3),alpha=0.04)

}
