% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Creates a histogram of sample means taken from a binomial}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{the sample size}

\item{iter}{the number of iterations to perform}

\item{p}{the probability of success}

\item{...}{stuff}
}
\value{
a histogram
}
\description{
Creates a histogram of sample means taken from a binomial
}
\examples{
mycltb(n=5,iter=10000,p=0.5)
}
