% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Creates a normal curve and displays the probability up to a certain point}
\usage{
myncurve(mu = 10, sigma = 2, a = 7)
}
\arguments{
\item{mu}{the mean}

\item{sigma}{the number to choose from}

\item{a}{the upper end of the area}
}
\value{
a curve
}
\description{
Creates a normal curve and displays the probability up to a certain point
}
\examples{
myncurve(mu=10,sigma=2,a=7)
}
