% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{A function determining number of tickets to be sold to optimize the overbooking problem}
\usage{
ntickets(N = 200, p = 0.95, gamma = 0.02)
}
\arguments{
\item{N}{the number of available seats on the flight}

\item{p}{the probability that a person shows up}

\item{gamma}{the probability of overbooking}
}
\value{
A discrete plot, a continuous curve, and a table of values
}
\description{
A function determining number of tickets to be sold to optimize the overbooking problem
}
\examples{
ntickets(N=400,p=0.94,gamma=0.02)

}
